% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGNMF.R
\name{divergenceCheck}
\alias{divergenceCheck}
\title{This function performs a divergence check, by comparing the current 
NMF sources with the initial ones. 3 divergence criteria are implemented.}
\usage{
divergenceCheck(W, W0)
}
\arguments{
\item{W}{Current NMF source matrix}

\item{W0}{Initial NMF source matrix}
}
\value{
Boolean value, indicating whether or not one of the 
divergence criteria has been reached
}
\description{
This function performs a divergence check, by comparing the current 
NMF sources with the initial ones. 3 divergence criteria are implemented.
}
\author{
Nicolas Sauwen
}
