% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HALSacc.R
\name{HALSupdt}
\alias{HALSupdt}
\title{Updating step for accelerated HALS NMF}
\usage{
HALSupdt(M, V, UtU, UtM, alpha, delta)
}
\arguments{
\item{M}{Input data matrix}

\item{V}{Factor matrix to be updated}

\item{UtU}{Product of the other transposed factor matrix with itself}

\item{UtM}{Product of the other transposed factor matrix with the input matrix}

\item{alpha}{Nonnegative parameter of the accelerated method}

\item{delta}{Convergence parameter}
}
\value{
Updated factor matrix V
}
\description{
Updating step for accelerated HALS NMF
}
\author{
Nicolas Sauwen
}
