% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imoverlay.R
\name{imoverlay}
\alias{imoverlay}
\title{Overlay a mask or a color scaled image on top of a background image}
\usage{
imoverlay(image, overlay, selectVect = NULL, color = c(0, 1, 0))
}
\arguments{
\item{image}{A matrix, background image}

\item{overlay}{A matrix, serving as the overlay mask or figure}

\item{selectVect}{A matrix (binary values), specifying which matrix elements are to be overlaid}

\item{color}{3-element vector, defining the RGB color to be used in case the overlay is a mask}
}
\description{
Overlay a mask or a color scaled image on top of a background image
}
\author{
Nicolas Sauwen
}
