% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importMatlabMriData.R
\name{importMatlabMriData}
\alias{importMatlabMriData}
\title{Interactive method to select Matlab (.mat) input dataset to perform 
(h)NMF analyses}
\usage{
importMatlabMriData()
}
\value{
List of input data attributes
}
\description{
Interactive method to select Matlab (.mat) input dataset to perform 
(h)NMF analyses
}
\author{
Nicolas Sauwen
}
