% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneLevelNMF.R
\name{initializeNMFModel}
\alias{initializeNMFModel}
\title{Initialize NMF model with initial spectral data}
\usage{
initializeNMFModel(X, initData = NULL)
}
\arguments{
\item{X}{input matrix}

\item{initData}{source or abundance matrix with initial values}
}
\description{
Initialize NMF model with initial spectral data
}
