% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGNMF.R
\name{nlsSubProb}
\alias{nlsSubProb}
\title{Algorithm for solving convex non-negative least squares subproblem using projected gradients}
\usage{
nlsSubProb(X, W, Hinit, tol, maxIter)
}
\arguments{
\item{X}{Input data matrix}

\item{W}{NMF basis matrix}

\item{Hinit}{Initial NMF coef matrix}

\item{tol}{Tolerance for a relative stopping condition}

\item{maxIter}{Maximum number of iterations}
}
\value{
List containing updated H matrix, its gradient and number of iterations
}
\description{
Algorithm for solving convex non-negative least squares subproblem using projected gradients
}
\author{
Nicolas Sauwen
}
