% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneLevelNMF.R
\name{preProcesInputData}
\alias{preProcesInputData}
\title{Condition input data matrix properly for NMF}
\usage{
preProcesInputData(X)
}
\arguments{
\item{X}{input matrix}
}
\value{
matrix with non-zero elements
}
\description{
Condition input data matrix properly for NMF
}
