% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residualNMF.R
\name{residualNMF}
\alias{residualNMF}
\title{Computation of relative NMF residual per observation}
\usage{
residualNMF(X, nmfFit)
}
\arguments{
\item{X}{Input data matrix, each column represents one observation}

\item{nmfFit}{NMF model fitted to the input data in X}
}
\value{
Relative residual per observation, returned as a vector
}
\description{
Computation of relative NMF residual per observation
}
\author{
nsauwen
}
