% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneLevelNMF.R
\name{scaleNMFResult}
\alias{scaleNMFResult}
\title{Apply fixed scaling to NMF model matrices by normalizing the basis vectors}
\usage{
scaleNMFResult(NMFResult)
}
\arguments{
\item{NMFResult}{Fitted NMF model}
}
\value{
NMFResult Rescaled NMF model
}
\description{
Apply fixed scaling to NMF model matrices by normalizing the basis vectors
}
\author{
Nicolas Sauwen
}
