% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewNMFResult.R
\name{viewNMFResult}
\alias{viewNMFResult}
\title{Visualization of (h)NMF results}
\usage{
viewNMFResult(nmfInput, nmfModel)
}
\arguments{
\item{nmfInput}{List with NMF input attributes}

\item{nmfModel}{NMF model (in accordance with NMF package definition)}
}
\description{
(h)NMF abundance maps are plotted overlaid on top of a background (input) image
}
\author{
Nicolas Sauwen
}
