% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hNMF.R
\name{hNMF}
\alias{hNMF}
\title{Hierarchical non-negative matrix factorization.}
\usage{
hNMF(nmfInput, nmfMethod = "HALSacc")
}
\arguments{
\item{nmfInput}{List with NMF input attributes}

\item{nmfMethod}{String referring to the NMF algorithm to be used.}
}
\value{
Resulting NMF model (in accordance with NMF package definition)
}
\description{
Hierarchical non-negative matrix factorization.
}
\author{
Nicolas Sauwen
}
