% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeSPA.R
\name{initializeSPA}
\alias{initializeSPA}
\title{The successive projection algorithm, a useful method for initializing the NMF source matrix}
\usage{
initializeSPA(data, nSources)
}
\arguments{
\item{data}{Input data matrix. The columns correspond to the data points, each row 
represents one feature}

\item{nSources}{Number of sources to be obtained}
}
\value{
Matrix with initialized sources as its columns
}
\description{
The successive projection algorithm, a useful method for initializing the NMF source matrix
}
\author{
Nicolas Sauwen
}
